refactor: code refactor; break js file on modules fix: delete extra files, extra fonts feat: add styles for all media, add functionality for state page.
