+++ index.html
@@ -6,8 +6,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link type="image/x-icon" href="/assets/images/favicon.ico" rel="icon">
-    <link rel="stylesheet" href="style.css" type="text/css">
-    <!-- <script src="/src/app.js" defer></script> -->
+    <link rel="stylesheet" href="style.css" media="screen and (min-width: 768px)" type="text/css">  
+    <link href="style320px.css" rel="stylesheet" media="screen and (max-width: 767px)" type="text/css">  
+    <script type="module" src="/dist/bundle.js" defer></script>
     <title>English for kids</title>
 </head>
 
@@ -56,15 +57,10 @@
         </main>
         <footer></footer>
     </div>
-<!-- <script src="rotate.js" defer></script> -->
-    <!-- <script src="/src/app.js" defer></script> -->
-    <!-- <script src="/src/variables.js" defer></script> -->
-    <!-- <script type="module" src="/class.js" defer></script> -->
-    <script type="module" src="/index.js" defer></script>
 
-    <!-- <script type="module" src="/localstorage.js" defer></script> -->
-    <script type="module" src="/change-mode.js" defer></script>
-    <script type="module" src="/game-mode.js" defer></script>
+    
+
+  
 </body>
 
 </html>
