+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 module.exports = {
   mode: 'development',
   devtool: 'eval',
-  entry: './src/app.js',
+  entry: './src/index.js',
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: 'bundle.js',
